% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wraper.R
\name{JTK}
\alias{JTK}
\title{Identify data with rhythms using JTK algorithm}
\usage{
JTK(
  data,
  timepoints,
  reps = 1,
  normal = FALSE,
  alt = FALSE,
  periods,
  interval,
  ampci = FALSE,
  conf = 0.8
)
}
\arguments{
\item{data}{expression matrix, each column represents a sample at
certain time, each row represents a gene/protein}

\item{timepoints}{total time points}

\item{reps}{replicates per time point}

\item{normal}{normal approximation}

\item{alt}{alternative distributions}

\item{periods}{rhythms of how many time points (i.e. between 5 and 7 time
points per cycle if periods = 5:7)}

\item{interval}{the number of time units (hours) between time points}

\item{ampci}{calculating amplitude confidence if TRUE, default is FALSE}

\item{conf}{confidence level}
}
\value{
a data frame showing information of genes/proteins rhythms.
"BH.Q" is adjusted p value using BH method, "ADJ.P" is the rhythms
significance p value, "PER" is periods (in time unit), "LAG" is the lag of
rhythms, "AMP" is the amplitudes.
}
\description{
This is a non-parametric algorithm for detecting rhythmic components
in genome-scale datasets.
}
\examples{
time = seq(1, 72, by = 3)
set.seed(123)
expr1 = 2 + 3 * cos(time/(6) - 4) + rnorm(length(time), 0, 0.2)
expr2 = 3 + 2 * cos(time/(3) - 3) + rnorm(length(time), 0, 0.2)
expr3 = 3 + 2 * cos(time/(6) - 3) + rnorm(length(time), 0, 2)
plot(expr3, col = "black", ylim = c(-1, 8))
lines(expr3, col = "black")
points(expr2, col = "blue")
lines(expr2, col = "blue")
points(expr1, col= "red")
lines(expr1, col = "red")

data = as.data.frame(rbind(expr1, expr2, expr3))

JTK(data, timepoints = 24, periods = 2:24, interval = 3)
}
