% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JTKversion4.R
\name{plotJTK}
\alias{plotJTK}
\title{Fit a cosine curve for rhythmic data}
\usage{
plotJTK(expr, Time, n = 6, periodHours = seq(1, 60, by = 0.5))
}
\arguments{
\item{expr}{numeric vector, expression of a gene/protein}

\item{Time}{numeric vector, all time points (time unit)}

\item{n}{number of best fitted plots}

\item{periodHours}{numeric vector, fit cosine curve (time unit)}
}
\value{
a ggplot of cosine curves with different periods.
}
\description{
Fit a cosine curve for rhythmic data
}
\examples{
time = seq(1, 72, by = 3)
set.seed(123)
expr1 = 2 + 3 * cos(time/(6) - 4) + rnorm(length(time), 0, 0.2)
expr2 = 3 + 2 * cos(time/(3) - 3) + rnorm(length(time), 0, 0.2)
expr3 = 3 + 2 * cos(time/(6) - 3) + rnorm(length(time), 0, 2)
plot(expr3, col = "black", ylim = c(-1, 8))
lines(expr3, col = "black")
points(expr2, col = "blue")
lines(expr2, col = "blue")
points(expr1, col= "red")
lines(expr1, col = "red")

data = as.data.frame(rbind(expr1, expr2, expr3))

result = JTK(data, timepoints = 24, periods = 2:24, interval = 3)
plotJTK(as.numeric(result[1, 6:ncol(result)]), time)
plotJTK(as.numeric(result[1, 6:ncol(result)]), time, 1, 42)
plotJTK(as.numeric(result[2, 6:ncol(result)]), time)
plotJTK(as.numeric(result[3, 6:ncol(result)]), time)
}
